--- src/main/java/com/kdt/springbootboardjpa/config/Config.java
@@ -1,8 +1,0 @@
-package com.kdt.springbootboardjpa.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-@Configuration
-@PropertySource("classpath:/application.yml")
-public class Config {}
